Change return type of calc_tree_size to long long
Updated the return type to ensure compatibility with larger tree sizes.

+++ rbtlib/rbtree.c
@@ -327,7 +327,7 @@ size_t calc_file_info_size(const FileInfo *fileInfo) {
 }
 
 // Function to calculate serialized size of the whole tree
-size_t calc_tree_size(const Node *node) {
+long long calc_tree_size(const Node *node) {
     if (node == NULL) {
         return 0;
     }
